// lib/types.ts
export interface PolicyData {
  url?: string;
  websiteType: string;
  otherType?: string;
  dataCollected: string[];
  collectionMethods: string[];
  dataUsage: string[];
  thirdPartySharing: string[];
  cookies: string;
  userRights: string[];
  businessName: string;
  contactEmail: string;
  country: string;
  state?: string;
}